+++ src/eu/qiou/aaf4k/reportings/etl/AccountingFrame.kt
@@ -163,7 +163,7 @@ class AccountingFrame(id: Int, name: String, val accounts: List<Account>) :
                     val arr = s.split("#")
                     val name = arr[1]
                     regIndent.find(s)?.groups!!.let {
-                        return Account(it.get(3)!!.value.toInt(),
+                        return Account(it.get(3)!!.value.toLong(),
                                 name, decimalPrecision = 2, value = 0,
                                 isStatistical = it.get(2)!!.value.length == 1,
                                 reportingType = if (types(arr) == null) t!!
@@ -176,7 +176,7 @@ class AccountingFrame(id: Int, name: String, val accounts: List<Account>) :
                     val arr = src.split("#")
                     val name = arr[1]
                     regIndent.find(src)?.groups!!.let {
-                        return Account(it.get(3)!!.value.toInt(), name, decimalPrecision = 2,
+                        return Account(it.get(3)!!.value.toLong(), name, decimalPrecision = 2,
                                 subAccounts = acc.toMutableList(),
                                 isStatistical = it.get(2)!!.value.length == 1,
                                 reportingType = if (types(arr) == null) t!! else types(arr)!!
