+++ src/eu/qiou/aaf4k/reportings/etl/DataLoader.kt
@@ -4,5 +4,5 @@ package eu.qiou.aaf4k.reportings.etl
  *  account id -> account value
  */
 interface DataLoader {
-    fun load(): MutableMap<Int, Double>
+    fun load(): MutableMap<Long, Double>
 }
